+++ noodles-vcf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * vcf/lazy/record: Implement `Clone` + `Eq` + `PartialEq`.
+
 ## 0.38.0 - 2023-08-31
 
 ### Added