+++ noodles-vcf/CHANGELOG.md
@@ -14,8 +14,8 @@
     When the input is VCF 4.2, this allows the `Child` or `Derived` field to
     act as the record ID in the value collection.
 
-  * vcf/header/parser/record/value/map/other: Only parse `Values` field in META
-    record.
+  * vcf/header/parser/record/value/map/other: Only parse `Values` field in VCF
+    4.3+ META records.
 
 [#201]: https://github.com/zaeleus/noodles/issues/201
 [#203]: https://github.com/zaeleus/noodles/issues/203